[INSTX-7939] Proper fix for modbase perf issues on TX2

See merge request machine-learning/dorado!1389

+++ tests/AdapterDetectorTest.cpp
@@ -395,7 +395,8 @@ CATCH_TEST_CASE(
     read->read_common.base_mod_probs[(flank_size * mod_alphabet.size()) + 1] = 235;  // 6mA
     read->read_common.num_trimmed_samples = 0;
 
-    auto records = read->read_common.extract_sam_lines(true /* emit moves */, 10, false);
+    auto records = read->read_common.extract_sam_lines(true /* emit moves */,
+                                                       static_cast<uint8_t>(10), false);
     BamPtr record_copy(bam_dup1(records[0].get()));
 
     auto client_info = std::make_shared<dorado::DefaultClientInfo>();
@@ -479,7 +480,7 @@ CATCH_TEST_CASE(
 
             CATCH_CHECK(read_common.num_trimmed_samples == uint64_t(additional_trimmed_samples));
 
-            auto bams = read_common.extract_sam_lines(0, 10, false);
+            auto bams = read_common.extract_sam_lines(0, static_cast<uint8_t>(10), false);
             auto& rec = bams[0];
             auto [mod_str, mod_probs] = dorado::utils::extract_modbase_info(rec.get());
         }
